requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-protected-media==0.0.3
docker==4.0.2
gunicorn==19.9.0
idna==2.8
pkg-resources==0.0.0
psycopg2==2.8.3
pyinotify==0.9.6
python-decouple==3.1
python-magic==0.4.15
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
websocket-client==0.56.0
